Adds wav streaming

+++ src/main/kotlin/audio/AudioTransform.kt
@@ -22,7 +22,7 @@ class AudioTransform(
     private val smoothingPeakCoeff = exp(-1.0 / (fps * 20.0)).toFloat() // 20000ms release-time
     private val smoothingEnergyCoeff = exp(-1.0 / (fps * 0.02)).toFloat() // 20ms release-time
 
-    private var position: Int = 0
+    private var position: Long = 0
 
     fun advance(targetTime: Double) {
         val targetFrame: Double = secondsToNumFrames(targetTime, format.sampleRate())
